requirements.txt
# Geocoding
geopy==2.4.1

# Data processing
pandas==2.3.3

# Retry logic
backoff==2.2.1

# Environment variables
python-dotenv==1.0.1

# Google Drive integration
google-api-python-client==2.121.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
